Add com.microsoft.hyperspace.index.rules.RuleUtils
